refactor: Phase 4 공용 UI 분리
- 범용 UI 컴포넌트들을 shared/components/common 에서 shared/ui 로 이동
  (Skeleton, LoadingSpinner, Pagination, BookmarkButton, LoginRequiredCard, ReportRequiredCard, ListStateUI, SelectDropdown, NotificationModal, Loading)
- tsconfig.json 에 @ui alias 추가
