+++ src/pages/my-project/my-project-page.tsx
@@ -1,5 +1,5 @@
 import { useAuth } from '@clerk/clerk-react';
-import LoginRequiredCard from '@components/common/LoginRequiredCard';
+import LoginRequiredCard from '@ui/LoginRequiredCard';
 import MyProjectTabs from './_components/MyProjectTabs';
 import { Outlet } from 'react-router-dom';
