+++ src/pages/report/_components/GithubRepoListSkeleton.tsx
@@ -1,4 +1,4 @@
-import Skeleton from '@components/common/Skeleton';
+import Skeleton from '@ui/Skeleton';
 
 type GithubRepoListSkeletonProps = {
   count?: number;
