+++ src/shared/components/common/ProjectBase.tsx
@@ -1,6 +1,6 @@
 import PersonIcon from '@assets/icons/person.svg?react';
 import DevineLogo from '@assets/images/Devine.svg';
-import BookmarkButton from '@components/common/BookmarkButton';
+import BookmarkButton from '@ui/BookmarkButton';
 import { badgeToneToClass } from '@t/badgeTone';
 import type { ProjectCardBaseProps } from '@t/project/ui';
 import type { KeyboardEvent } from 'react';
