+++ src/shared/components/common/RecommendDeveloperCardSkeleton.tsx
@@ -1,4 +1,4 @@
-import Skeleton from '@components/common/Skeleton';
+import Skeleton from '@ui/Skeleton';
 
 export default function RecommendDeveloperCardSkeleton() {
   return (
