+++ src/shared/ui/LoginRequiredCard.tsx
