+++ src/shared/ui/ReportRequiredCard.tsx
